% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.bHP.R
\name{BIC.bHP}
\alias{BIC.bHP}
\title{all in one function of iterated HP-filter}
\usage{
\method{BIC}{bHP}(x)
}
\arguments{
\item{x}{the data you want to conduct HP-filter}
}
\value{
cycle component, iterated number, p-value .
}
\description{
all in one function of iterated HP-filter
}
\examples{
lam <- 100 # tuning parameter for the annaul data

# raw HP filter
bx_HP <- BoostedHP(x, lambda = lam, iter= FALSE)$trend

# by BIC
bx_BIC <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "BIC")

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "adf", sig_p = 0.050)

# summarize the outcome
outcome <- cbind(IRE, bx_HP$trend, bx_BIC$trend, bx_ADF$trend)

matplot(outcome, type = "l", ylab = "", lwd = rep(2,4))


}
