% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{bye}
\alias{bye}
\title{bye function}
\usage{
bye(x = "Analyx")
}
\arguments{
\item{x}{A character. The name of who you want to say hi.}

\item{y}{Logical. True if you want a formal salute.}
}
\value{
A string containing the salute
}
\description{
bye function
}
\examples{
bye()
bye("Adolfo")
bye("Adolfo", formal = FALSE)
}

