% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasR-package.R
\docType{package}
\name{debiasR-package}
\alias{debiasR}
\alias{debiasR-package}
\title{debiasR: Bias Correction for Mobile-Phone-Derived Mobility Data}
\description{
Tools to correct biases in origin-destination mobility data derived from mobile phone sources. Implements stable deterministic bias-adjustment methods plus Bayesian multilevel adjustment, including coverage-offset true-flow models for observed and complete-grid prediction and latent true-flow models for repeated source/time workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://de-bias.github.io/debiasR/}
  \item \url{https://github.com/de-bias/debiasR}
  \item Report bugs at \url{https://github.com/de-bias/debiasR/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rowe \email{fcorowe@liverpool.ac.uk} (\href{https://orcid.org/0000-0003-4137-0246}{ORCID})

Authors:
\itemize{
  \item Francisco Rowe \email{fcorowe@liverpool.ac.uk} (\href{https://orcid.org/0000-0003-4137-0246}{ORCID})
  \item Carmen Cabrera \email{c.cabrera@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-2732-6436}{ORCID})
}

}
\keyword{internal}
