% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world function}
\usage{
hello(x = "Analyx", formal = TRUE)
}
\arguments{
\item{x}{A character. The name of who you want to say hi.}

\item{y}{Logical. True if you want a formal salute.}
}
\value{
A string containing the salute
}
\description{
Hello world function
}
\examples{
hello()
hello("Adolfo")
hello("Adolfo", formal = FALSE)
}

