% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Hello world function 2}
\usage{
hello2(x = "Analyx", formal = TRUE, bye = TRUE)
}
\arguments{
\item{x}{A character. The name of who you want to say hi.}

\item{y}{Logical. True if you want a formal salute.}
}
\value{
A string containing the salute
}
\description{
Hello world function 2
}
\examples{
hello2()
hello2("Adolfo")
hello2("Adolfo", formal = FALSE)
}

