% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{MM}
\alias{MM}
\title{MM}
\usage{
MM(
  x,
  y = c(),
  N = Inf,
  setting = "exp",
  alpha = 0,
  epsilon = 0,
  penalty = FALSE,
  auto_gcm = TRUE,
  auto_bisection = FALSE,
  direct_c = TRUE
)
}
\arguments{
\item{x}{observed biased data}

\item{y}{observed unbiased data, can be NULL}

\item{N}{population size, default to infinity}

\item{setting}{assumed null distribution, default to exponential}

\item{alpha}{penalty term for MM estimation}

\item{epsilon}{assumed minimum observation probability}

\item{penalty}{(ignore this)}

\item{auto_gcm}{if TRUE, use the "gcmlcm" function from package "fdrtool" to perform fast GCM computation; else use R based functions}

\item{auto_bisection}{if TRUE, use "optim" or "optimize" function to perform bisection algorithm, else use manually written functions}

\item{direct_c}{if TRUE, directly use the C function "isomean" from package "fdrtools" instead of using "gcmlcm" function}
}
\value{
estimation result
}
\description{
Perform MM-estimation
}
\examples{
MM(rexp(30),setting='exp',alpha=0.5,epsilon=0.2)
MM(rnorm(30),rnorm(10),setting='norm',alpha=0.5,epsilon=0.2)
}
