% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bHP.R
\name{plot.bHP}
\alias{plot.bHP}
\title{all in one function of iterated HP-filter}
\usage{
\method{plot}{bHP}(x, plot_type = "dynamic", interval_t = 0.3,
  ylab = "", col_raw = "#2D5375", col_trend_h = "#FBB545",
  col_trend_f = "red", col_pvalue_BIC = "red", raw_alpha = 255,
  trend_h_alpha = 75, trend_f_alpha = 255, pvalue_BIC_alpha = 255,
  legend_location = "upleft", iteration_location = "downright",
  cex_text = 1.7, cex_legend = 1.5, main = paste0("Figure of ",
  x$test_type, " bHP (", plot_type, ")"))
}
\arguments{
\item{x}{the data you want to conduct HP-filter}

\item{lambda}{the turning parameter}

\item{iter}{logical parameter, TRUE is to conduct iterated HP-filter, FALSE is not}

\item{test_type}{the type for creterion}

\item{sig_p}{significant p-value}

\item{Max_Iter}{maximum iterated time}
}
\value{
cycle component, iterated number, p-value .
}
\description{
all in one function of iterated HP-filter
}
\examples{
lam <- 100 # tuning parameter for the annaul data

# raw HP filter
bx_HP <- BoostedHP(x, lambda = lam, iter= FALSE)$trend
plot(bx_HP)

# by BIC
bx_BIC <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "BIC")

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "adf", sig_p = 0.050)

# summarize the outcome
outcome <- cbind(IRE, bx_HP$trend, bx_BIC$trend, bx_ADF$trend)



}
