% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_of_P_test.R
\name{power_of_P_test}
\alias{power_of_P_test}
\title{Power of the P-test}
\usage{
power_of_P_test(setting = "ux", N, C)
}
\arguments{
\item{setting}{the distribution of null and alternative}

\item{N}{sample size, can be a vector}

\item{C}{constant parameter, can be a vector}
}
\value{
power table
}
\description{
Compute the power of the P-test under specific settings
}
\examples{
power_of_P_test('ux',c(10,20,30),c(0.2,0.3))
}
