% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvt.isomean.R
\name{pvt.isoMean}
\alias{pvt.isoMean}
\title{pvt.isoMean}
\usage{
pvt.isoMean(y, w)
}
\arguments{
\item{y}{y values}

\item{w}{weights}
}
\value{
estimated gcm
}
\description{
pvt.isoMean
}
\examples{
pvt.isoMean(rnorm(10),rep(1,10))
}
