% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnG_UL.R
\name{RnG_UL}
\alias{RnG_UL}
\title{Title}
\usage{
RnG_UL(n, m = 0, N = -Inf, rnull = rexp, bias_type = "971", b = 0, ...)
}
\arguments{
\item{n}{number of biased sample}

\item{m}{number of unbiased sample}

\item{N}{population sizze}

\item{rnull}{function to generate samples from null distribution, e.g., rexp, rnorm}

\item{bias_type}{format of bias function}

\item{b}{a parameter in bias function, usually larger b represents larger bias}

\item{...}{additional arguments used in 'rnull', e.g., mean, sd}
}
\value{
a sample of biased (+unbiased) sample
}
\description{
Title
}
\examples{
RnG_UL(n=10,m=2,rnull=rexp,bias_type='971',b=3)
}
