% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2-package.R
\docType{package}
\name{test2-package}
\alias{test2}
\alias{test2-package}
\title{test2: This Package Contains Data From A Salivary Metabolome Study}
\description{
An example of minimal package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stanciua19/test2}
  \item Report bugs at \url{https://github.com/stanciua19/test2/issues}
}

}
\author{
\strong{Maintainer}: Ana Stanciu \email{stanciua@mcmaster.ca}

}
\keyword{internal}
