% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2-package.R
\docType{package}
\name{test2}
\alias{test2}
\title{test2: This is a package with minimum example of package creation.
This package will contain data from a salivary metabolomic study and will be used in the creation of a minimal package.}
\references{
\{https://github.com/stanciua19/test2}
}
\author{
Ana Stanciu, Chemical Biology, McMaster University
}
