% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDmodiflyggplot.R
\name{zstatmodifyplot}
\alias{zstatmodifyplot}
\title{增强 ggplot2 图形美化函数}
\usage{
zstatmodifyplot(
  p,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  log_y = FALSE,
  theme_background = c("white", "gray", "minimal", "none"),
  show_grid = TRUE,
  line_size = 1,
  line_type = "solid",
  show_CI = TRUE,
  color_palette = NULL,
  group_color_scheme = c("none", "groupcolor1", "groupcolor2", "groupcolor3"),
  panel_border = c("none", "black", "gray", "custom"),
  border_color = "black",
  border_size = 0.5,
  legend_position = "right",
  legend_direction = "vertical",
  text_size = 12,
  axis_text_size = 10,
  legend_text_size = 10,
  axis_line = TRUE,
  axis_line_sides = c("left", "bottom"),
  axis_line_size = 0.5,
  axis_line_color = "black",
  facet_spacing = 1,
  plot_margin = c(10, 10, 10, 10),
  font_family = "Arial",
  xlab_margin = 10,
  ylab_margin = 10,
  legend_title = NULL,
  legend_labels = NULL,
  x_angle = NULL
)
}
\arguments{
\item{p}{各画图函数生成的ggplot 图形对象}

\item{title}{主标题}

\item{subtitle}{副标题}

\item{xlab}{x轴标签}

\item{ylab}{y轴标签}

\item{xlim}{x轴范围（可选）}

\item{ylim}{y轴范围（可选）}

\item{log_y}{是否对y轴取log10}

\item{theme_background}{图形主题，"white"（默认）、"gray"、"minimal"、"none"，推荐使用默认}

\item{show_grid}{是否显示网格线，默认TRUE}

\item{line_size}{主线条粗细，默认1}

\item{line_type}{主线条类型，默认"solid"}

\item{show_CI}{是否显示置信区间图层}

\item{color_palette}{手动设置颜色，如为命名向量}

\item{group_color_scheme}{内置颜色方案（"none", "groupcolor1", "groupcolor2", "groupcolor3"），若你在原有画图函数中已经配好颜色，请在此处设置group_color_scheme="none"}

\item{panel_border}{是否绘制边框，默认"none"，可设为"black"、"gray"、"custom"，推荐使用默认}

\item{border_color}{自定义边框颜色，仅在 panel_border = "custom" 时有效}

\item{border_size}{边框线宽，默认0.5}

\item{legend_position}{图例位置（"right", "bottom", "top", "left"）}

\item{legend_direction}{图例排列方向（"vertical", "horizontal"）}

\item{text_size}{全局文字大小}

\item{axis_text_size}{坐标轴文字大小}

\item{legend_text_size}{图例文字大小}

\item{axis_line}{是否绘制轴线}

\item{axis_line_sides}{轴线显示方向，向量形式如 c("left", "bottom")}

\item{axis_line_size}{坐标轴线宽}

\item{axis_line_color}{坐标轴线颜色}

\item{facet_spacing}{分面之间间距,默认为1}

\item{plot_margin}{图形边距，单位为points,默认plot_margin = c(10, 10, 10, 10)}

\item{font_family}{字体族，默认"Arial"}

\item{xlab_margin}{x轴标签与轴的间距}

\item{ylab_margin}{y轴标签与轴的间距}

\item{legend_title}{图例标题}

\item{legend_labels}{图例标签}

\item{x_angle}{x轴文字旋转角度（例如45）,默认x_angle = NULL}
}
\value{
一个增强美化后的 ggplot 图形对象
}
\description{
用于对已有的 ggplot 图形对象进行美化和格式增强，包括主题、边框、坐标轴、图例、线条等设置。
}
