% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstatsAPC.R
\name{zstatsapc}
\alias{zstatsapc}
\title{年龄周期队列(Age-Period-Cohort,APC)模型}
\usage{
zstatsapc(
  data,
  startyear = 1992,
  endyear = 2021,
  reference_age = NaN,
  reference_year = NaN,
  measure_name = "your measure",
  cause_name = "your cause",
  sex_name = "your sex",
  location_name = "your location",
  rei_name = NULL
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框)，包含measure,location,sex,age,cause,rei(如果含有),metric,year,val,lower,upper等列。}

\item{startyear}{起始年份，默认1992}

\item{endyear}{截止年份，默认2021}

\item{reference_age}{参考年龄段，默认NaN表示自动}

\item{reference_year}{参考年份，默认NaN表示自动}

\item{measure_name}{指标名称，如"Incidence"}

\item{cause_name}{病因名称}

\item{sex_name}{性别}

\item{location_name}{地理位置}

\item{rei_name}{风险因素，可为NULL}
}
\value{
一个包含 APC 各种输出的 list，包含NetDrift、AgeDeviations等结果数据框
}
\description{
对GBD数据执行APC分析，支持包含或不包含风险因素(rei)变量。
}
