% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDASR.R
\name{zstatsASR}
\alias{zstatsASR}
\title{计算不含有风险因素变量(rei)特定年龄的年龄标准化率(Age_standardized Rate, ASR)}
\usage{
zstatsASR(data, startage = 5, endage = 70, CI = FALSE)
}
\arguments{
\item{data}{不含有风险因素变量(rei)的数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)}

\item{startage}{起始年龄，若起始年龄是0 to 4,填入=0；若是5 to 9，填入=5，以此类推。}

\item{endage}{结束年龄,若结束年龄是90 to 94填入=90；若是95 plus，填入=95，以此类推。}

\item{CI}{是否计算置信区间,默认FALSE}
}
\value{
包含Age-standardized和All ages的Rate数据
}
\description{
计算不含有风险因素变量(rei)特定年龄的年龄标准化率(Age_standardized Rate, ASR)
}
