% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDASRrei.R
\name{zstatsASR_rei}
\alias{zstatsASR_rei}
\title{计算含有风险因素变量(rei)特定年龄的年龄标准化率(Age_standardized Rate, ASR)}
\usage{
zstatsASR_rei(
  rei_data,
  GBDage_recal_result = NULL,
  startage = 5,
  endage = 70,
  CI = FALSE
)
}
\arguments{
\item{rei_data}{含有风险因素变量（rei）的数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)}

\item{GBDage_recal_result}{默认为NULL，代表不计算人群归因分数PAF，如果需计算PAF，请提供zstatsASR()函数计算结果代入}

\item{startage}{起始年龄，若起始年龄是0 to 4,填入=0；若是5 to 9，填入=5，以此类推。}

\item{endage}{结束年龄,若结束年龄是90 to 94填入=90；若是95 plus，填入=95，以此类推。}

\item{CI}{是否计算置信区间,默认FALSE}
}
\value{
包含Age-standardized和All ages的Rate数据
}
\description{
计算含有风险因素变量(rei)特定年龄的年龄标准化率(Age_standardized Rate, ASR)
}
