% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDBAPC.R
\name{zstatsBAPC}
\alias{zstatsBAPC}
\title{贝叶斯年龄-周期-队列模型预测(Bayesian APC Prediction,BAPC)}
\usage{
zstatsBAPC(
  data,
  measure_name,
  cause_name,
  location_name,
  rei_name = NULL,
  By_sex = T,
  predyear = 2036,
  full_age_adjusted = F,
  rate_lessen = NULL
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框)，，包含measure,location,sex,age,cause,rei(如果含有),metric,year,val,lower,upper等列。}

\item{measure_name}{指标名称（如 "Incidence"、"Deaths"）。}

\item{cause_name}{病因名称。}

\item{location_name}{地区名称。}

\item{rei_name}{可选，风险因素名称。}

\item{By_sex}{是否输出分性别预测结果,默认TRUE(如果你主要关注性别差异，并只把 Both 作为一个合并参考 → 使用默认 By_sex = TRUE 即可),如果你有完整的 Both 数据，并希望建模其独立的趋势变化 → 请设定 By_sex = FALSE。}

\item{predyear}{预测截至年份,默认为2036年。}

\item{full_age_adjusted}{是否包含完整年龄段,默认FALSE。}

\item{rate_lessen}{可选，对预测结果进行缩放调整的系数，默认是NULL，如果设置缩放倍数，必须为数值。}
}
\value{
一个包含多个预测结果数据框的列表，包括：\cr
\itemize{
\item \code{all_age_projection}
\item \code{crude_rate}
\item \code{ASR}
\item \code{Age_standardized_projection}
\item \code{age_specific_rate}
\item \code{age_specific_projection}
}
}
\description{
使用 BAPC 方法对指定 GBD 数据进行发病/死亡数预测，并输出多种结果（ASR、Crude rate、Age-specific 等），支持包含或不包含风险因素(rei)变量。
}
