% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDdriftplot.R
\name{zstatsdrift}
\alias{zstatsdrift}
\title{绘制 LocalDrift 与 NetDrift 年龄相关趋势图}
\usage{
zstatsdrift(
  data,
  CI = TRUE,
  group_name = "your group",
  line_size = 1,
  measure_name = "your measure",
  cause_name = "your cause",
  sex_name = "your sex",
  location_name = "your location",
  rei_name = NULL
)
}
\arguments{
\item{data}{包含 LocalDrifts 和 NetDrift 的 list，通常为 zstatsapc() 函数的结果。}

\item{CI}{是否显示置信区间，默认 TRUE。}

\item{group_name}{分组变量名称，例如 "sex" 或 "location"。}

\item{line_size}{线条粗细，默认 1。}

\item{measure_name}{测量指标名称，如“Deaths"。}

\item{cause_name}{疾病名称。}

\item{sex_name}{性别。}

\item{location_name}{地点（国家或地区名称）。}

\item{rei_name}{风险因素（可选）。}
}
\value{
ggplot2 图形对象
}
\description{
根据 APC 模型函数生成的结果列表绘制 LocalDrift 与 NetDrift 年龄相关趋势图
}
