% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDEAPC.R
\name{zstatsEAPC}
\alias{zstatsEAPC}
\title{计算估计年度百分比变化EAPC（Estimated Annual Percentage Change）}
\usage{
zstatsEAPC(data, rei = FALSE, EAPC_95CI = TRUE, digits = 2, sep = " to ")
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)，要求含有列：location, measure, sex, cause, age, year, val, metric（和 rei，如 \code{rei = TRUE}）}

\item{rei}{数据框中是否含风险因素变量（rei），默认rei = FALSE，若存在rei，则rei = TRUE，否则rei = FALSE。}

\item{EAPC_95CI}{是否合并输出带格式的 EAPC + 置信区间（TRUE/FALSE）}

\item{digits}{保留小数位数，默认为 2}

\item{sep}{置信区间连接符号，默认为sep = " to ",其他可选择如 \code{" to "}、\code{"-"} 等}
}
\value{
一个包含 EAPC 和其 95\% CI 的数据框
}
\description{
使用线性回归模型对 GBD 率性指标进行趋势分析，返回 EAPC 和其 95\% 置信区间
}
