% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDFactor.R
\name{zstatsFactor}
\alias{zstatsFactor}
\title{将GBD变量转换成有序因子}
\usage{
zstatsFactor(
  data,
  location_order = FALSE,
  measure_order = FALSE,
  cause_order = FALSE,
  sex_order = FALSE,
  age_order = TRUE,
  rei_order = FALSE
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)}

\item{location_order}{是否处理 location 排序，默认FALSE不处理}

\item{measure_order}{是否处理 measure 排序，默认FALSE不处理}

\item{cause_order}{是否处理 cause 排序，默认FALSE不处理}

\item{sex_order}{是否处理 sex 排序，默认FALSE不处理}

\item{age_order}{是否处理 age 排序，默认TRUE处理}

\item{rei_order}{是否处理 rei 排序或自定义顺序，默认FALSE不处理}
}
\value{
一个变量有序化后的数据框
}
\description{
因子化数据框中的各变量，使得后续分析更有序化
}
