% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDfilter.R
\name{zstatsfilter}
\alias{zstatsfilter}
\title{GBD 数据快速筛选函数}
\usage{
zstatsfilter(
  data,
  measure = NULL,
  sex = NULL,
  age = NULL,
  location = NULL,
  metric = NULL,
  rei = NULL,
  cause = NULL,
  year = NULL,
  year_range = NULL,
  exclude_location = FALSE
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框)}

\item{measure}{可选，指标类型，如 "1" 或 "Incidence",例如measure=1表示仅筛选Incidence,measure=c("1","2")表示筛选Incidence和Prevalence（下列各指标筛选与此相同）,measure各数字的含义c("1" = "Incidence", "2" = "Prevalence", "3" = "Deaths", "4" = "DALYs (Disability-Adjusted Life Years)","5" = "YLDs (Years Lived with Disability)", "6" = "YLLs (Years of Life Lost)", "7" = "Summary exposure value")}

\item{sex}{可选，性别，如 "1" 或 "Male",sex各数字含义c("1" = "Male", "2" = "Female", "3" = "Both")}

\item{age}{可选，年龄组编号或名称，如 "1" 或 "0 to 4",age各数字含义c("1" = "Age-standardized", "2" = "All ages", "3" = 全年龄段0-95+岁, "4" = 儿童0-14岁, "5" = 绝经前/育龄妇女/青少年和成年人15-49岁, "6" = 儿童和青少年0-24岁, "7" = 青少年和年轻人10-24岁, "8" = 绝经后50-95+岁, "9" = 老年人60岁以上)}

\item{location}{可选，地区编号或名称，如 "1" 或 "South Asia",location各数字含义c("1" = 7个超级地区, "2" = 5个SDI地区, "3" = 21个GBD地区)}

\item{metric}{可选，数值类型，如 "1" 或 "Number",metric各数字含义c("1" = "Number", "2" = "Rate", "3" = "Percent")}

\item{rei}{可选，风险因素名}

\item{cause}{可选，疾病名称}

\item{year}{可选，年份（向量）比如year =2021表示仅筛选2021年数据,year=1990:2021表示筛选1990-2021年数据}

\item{year_range}{可选，年份范围，格式为 c(起始年, 结束年)}

\item{exclude_location}{是否从结果中排除指定地区(即location参数筛选的地区)，默认为 FALSE}
}
\value{
筛选后的数据框
}
\description{
用于按条件快速筛选 GBD 数据表(如measure、location、sex、age、cause、rei、metric、year等)，支持通过编号映射和组合分类过滤。
}
