% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDfrontier.R
\name{zstatsFrontier}
\alias{zstatsFrontier}
\title{前沿分析（Frontier Analysis）}
\usage{
zstatsFrontier(
  data,
  sex_name = "Both",
  cause_name,
  measure_name = "Deaths",
  rei_name = NULL,
  age_name = "age-stanardized",
  boot = 100,
  cpu_num = "half"
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)}

\item{sex_name}{性别（默认 "Both"）}

\item{cause_name}{病因}

\item{measure_name}{指标名称，默认 "Deaths"}

\item{rei_name}{风险因素，可选，默认 NULL}

\item{age_name}{年龄，默认 "age-stanardized"}

\item{boot}{自助法抽样次数，默认 100}

\item{cpu_num}{并行 CPU 数量，"half"、"max" 或整数，默认 "half"}
}
\value{
含前沿率、效率差异、趋势等信息的数据框
}
\description{
对 GBD 数据进行国家-年份层面的率与前沿效率分析
}
