% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDplotline.R
\name{zstatsline}
\alias{zstatsline}
\title{绘制折线图（按年龄或年份）}
\usage{
zstatsline(
  data,
  x_axis = "age",
  group = "sex",
  ASR = TRUE,
  groupcolor = "a",
  group_colors = NULL,
  show_CI = TRUE,
  ylabel = NULL,
  xlab = NULL,
  legend_title_group = NULL,
  x_breaks = NULL
)
}
\arguments{
\item{data}{数据框，包含 val、metric、age、year、sex、location、lower、upper 等变量}

\item{x_axis}{X轴变量，可为 "age" 或 "year"}

\item{group}{分组变量，默认为 "sex"}

\item{ASR}{是否使用 Age-standardized（仅当 x_axis = "year" 时有效）}

\item{groupcolor}{预设颜色方案，可为 "a"、"b"、"c"}

\item{group_colors}{自定义颜色向量（命名向量）}

\item{show_CI}{是否显示置信区间，默认 TRUE}

\item{ylabel, }{xlab, legend_title_group 自定义坐标轴/图例名称。设为 FALSE 表示不显示。}

\item{x_breaks}{设置 x 轴刻度。若为 NULL，则自动生成；否则手动设定。}
}
\value{
ggplot2 图形对象
}
\description{
绘制折线图（按年龄或年份）
}
