% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDMAP.R
\name{zstatsMap}
\alias{zstatsMap}
\title{世界地图可视化函数}
\usage{
zstatsMap(
  data,
  variable,
  discrete_colors = NULL,
  color_group = "a",
  guide_name = NULL,
  continuous_colors = NULL,
  cont_color_group = "a",
  show_na = TRUE,
  show_small_map = TRUE
)
}
\arguments{
\item{data}{包含 location 和 val 的数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)。}

\item{variable}{要映射的变量名(字符串),如variable = "val"。}

\item{discrete_colors}{离散变量自定义配色(向量)如discrete_colors = c("#EDF8FB", "#CCECE6", "#99D8C9", "#66C2A4", "#41AE76", "#238B45", "#005824")或discrete_colors = c("<0.27" = "#EDF8FB", "0.27 to <0.37" = "#CCECE6", "0.37 to <0.45" = "#99D8C9", "0.45 to <0.52" = "#66C2A4", "0.52 to <0.66" = "#41AE76", "0.66 to <0.94" = "#238B45", "0.94 to <2.2" = "#005824")。}

\item{color_group}{离散变量配色方案("a", "b", "c"等)。}

\item{guide_name}{图例标题,如guide_name = "ASMR(per 100,000)"。}

\item{continuous_colors}{连续变量自定义配色（向量），优先于 cont_color_group，如continuous_colors = c("#D73027", "#F46D43", "#FDAE61", "#FEE08B", "#D9EF8B", "#A6D96A", "#66BD63", "#1A9850")}

\item{cont_color_group}{连续变量配色方案("a", "b", "c"等)。}

\item{show_na}{是否显示缺失国家（如果显示确实则默认显示为灰色）。}

\item{show_small_map}{是否显示小地图区域。}
}
\value{
返回 ggplot 对象，含主图和小地图。
}
\description{
根据变量类型（连续或离散）自动调用适当的地图绘图函数。
例如，绘制年龄标准化发病率的世界地图，则需筛选measure(Incidence),location(204个国家),sex(仅含一种),age(Age_standardized),cause(仅含一种),rei(若有，仅含一种),metric(Rate),year(仅含一年),val
如果要将val进行分则，则使用zstatsquantile()函数进行分组，yourdata$val2 <- zstatsquantile(data= yourdata$val, n = 5)
}
