% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDfrontierplot.R
\name{zstatsplot_frontier}
\alias{zstatsplot_frontier}
\title{绘制 Frontier 前沿图}
\usage{
zstatsplot_frontier(
  frontier_result,
  smooth_span = 0.3,
  type = c("all years", "single year"),
  high_SDI = 0.85,
  low_SDI = 0.5
)
}
\arguments{
\item{frontier_result}{使用 Frontier 函数得到的结果数据框。}

\item{smooth_span}{LOESS 平滑参数，默认 0.3。}

\item{type}{图形类型，"all years" 或 "single year"。}

\item{high_SDI}{高 SDI 阈值，用于标注国家，默认 0.85。}

\item{low_SDI}{低 SDI 阈值，用于标注国家，默认 0.5。}
}
\value{
Frontier 前沿图
}
\description{
对 Frontier 分析结果进行可视化，展示高/低 SDI 国家和效率边界
}
