% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDpyramidplot.R
\name{zstatsplotpyramid}
\alias{zstatsplotpyramid}
\title{GBD 金字塔图绘制函数}
\usage{
zstatsplotpyramid(
  data,
  group = "location",
  x_axis = "age",
  percent = FALSE,
  location_name = "Global",
  cause_name = "Inflammatory bowel disease",
  measure_name = "Deaths",
  year_name = 2021,
  age_name = c("0 to 4", "5 to 9", "10 to 14", "15 to 19", "20 to 24", "25 to 29",
    "30 to 34", "35 to 39", "40 to 44", "45 to 49", "50 to 54", "55 to 59", "60 to 64",
    "65 to 69", "70 to 74", "75 to 79", "80 to 84", "85 to 89", "90 to 94", "95 plus"),
  metric_name = "Number",
  rei_name = NULL,
  color_scheme = "groupcolor1",
  show_value = TRUE,
  value_text_size = 2
)
}
\arguments{
\item{data}{数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)，需包含 sex、val、metric、year、measure、location、age、cause、rei(如果数据中含有)等变量}

\item{group}{分组变量（"age"、"location"、"cause"、"rei"、"year"），默认 "location"}

\item{x_axis}{横轴变量（默认与 group 相同）}

\item{percent}{是否转换为百分比显示，默认 FALSE}

\item{location_name}{地点名称，如 "Global",如要比较global和china，则location_name = c("Global","China")}

\item{cause_name}{疾病名称，如 "Inflammatory bowel disease"}

\item{measure_name}{指标名称，如 "Deaths"}

\item{year_name}{年份（数值），如 2021,如果以year为分组比较1990年和2021年，则year_name = c(1990, 2021)}

\item{age_name}{年龄组（字符向量），默认为全年龄段}

\item{metric_name}{指标单位（如 "Number"、"Rate"、"Percent"）}

\item{rei_name}{风险因素名称，若不涉及可设为 NULL}

\item{color_scheme}{内置颜色方案（groupcolor1/2/3/sexcolor），也可为颜色向量，即可手动设置颜色}

\item{show_value}{是否在图中显示数值标签，默认 TRUE}

\item{value_text_size}{控制图中显示数值标签字体大小，默认为2}
}
\value{
ggplot 对象，可直接打印或保存
}
\description{
用于绘制性别分层的 GBD 疾病负担金字塔图（Male 为正，Female 为负）。支持按不同分组维度（age/location/cause/rei）绘制，展示绝对值或百分比。
}
