% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDsdiASR.R
\name{zstatsplotSDIASR}
\alias{zstatsplotSDIASR}
\title{SDI 与 ASR 的相关性图}
\usage{
zstatsplotSDIASR(
  data,
  span_val = 0.5,
  show_se = FALSE,
  show_pvalue = TRUE,
  highlight_location = NULL,
  highlight_color = NULL,
  standard = c("Age-standardized", "All ages"),
  method = c("pearson", "kendall", "spearman"),
  fig_type = c("1", "2")
)
}
\arguments{
\item{data}{包含 location、year、age、val 等变量的 GBD 数据框(zstatsreadGBD()函数生成的GBD数据框或使用zstatsfilter()函数筛选后生成的数据框)}

\item{span_val}{loess 平滑曲线的 span 值，默认 0.5}

\item{show_se}{是否显示 loess 的置信区间（标准误），默认 FALSE}

\item{show_pvalue}{是否在图中添加相关系数 R 和 p 值，默认 TRUE}

\item{highlight_location}{指定需要高亮显示的国家名称（字符向量），默认 NULL}

\item{highlight_color}{若设置高亮国家，需提供命名颜色向量，如 c("China" = "red", "India" = "blue")}

\item{standard}{选择使用的年龄标准（"Age-standardized" 或 "All ages"）}

\item{method}{相关系数类型，"pearson"、"kendall"、"spearman"}

\item{fig_type}{图形类型，1或2,当fig_type=1时表示绘制的是1990-2021年的21个GBD地区的疾病负担与SDI之间相关性，当fig_type=2时表示绘制的是2021年的204个国家的疾病负担与SDI之间相关性}
}
\value{
ggplot 图形对象
}
\description{
用于展示社会发展指数（SDI）与疾病负担指标（ASR、死亡率等）之间的关系。支持选择年份范围或单一年份，设置平滑参数、显示置信区间和相关系数。
}
