% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDpredictplot2.R
\name{zstatspredictASR}
\alias{zstatspredictASR}
\title{年龄标准化率预测趋势图（ASR）}
\usage{
zstatspredictASR(
  data,
  CI = T,
  predict_start = 2022,
  group_name = "your group",
  location_name = "your location",
  measure_name = "your measure",
  cause_name = "your cause",
  rei_name = NULL,
  sex_name = "your sex"
)
}
\arguments{
\item{data}{\code{zstatsBAPC()} 函数返回的结果列表，必须包含 \code{ASR} 数据框。}

\item{CI}{是否绘制置信区间（默认 TRUE）。}

\item{predict_start}{预测起始年份，默认 2022。}

\item{group_name}{分组变量列名（如 "sex", "location", "cause" 等）。}

\item{location_name}{地点名称。}

\item{measure_name}{指标名称。}

\item{cause_name}{病因名称。}

\item{rei_name}{风险因素名称，默认为 NULL。}

\item{sex_name}{性别(如 "Male", "Female", "Both"),如分组变量是"sex",则sex_name = c("Male", "Female", "Both")(任意组合)，如果分组变量不为"sex",则sex_name = "Male"(仅举一例)。}
}
\value{
ggplot2 图形对象
}
\description{
用于绘制 \code{zstatsBAPC()} 输出结果中的 ASR（Age-Standardized Rate）预测趋势图，支持显示置信区间。
}
