% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDpredictplot.R
\name{zstatspredictNumberASR}
\alias{zstatspredictNumberASR}
\title{绘制 GBD 预测结果双轴图}
\usage{
zstatspredictNumberASR(
  data,
  ratio = "auto",
  CI = T,
  predict_start = 2022,
  group_name = "your group",
  location_name = "your location",
  measure_name = "your measure",
  cause_name = "your cause",
  rei_name = NULL,
  sex_name = "your sex"
)
}
\arguments{
\item{data}{\code{zstatsBAPC()} 返回的列表，包含 \code{ASR} 和 \code{all_age_projection}。}

\item{ratio}{控制ASR折线图的缩放比例，默认为 \code{"auto"} 自动缩放,也可设为数值(相对于"auto"比例的倍数，如ratio = 0.8,即表示缩放比例为"auto"时的比例的0.8倍)}

\item{CI}{是否显示置信区间(默认 TRUE)。}

\item{predict_start}{预测起始年份(默认 2022)。}

\item{group_name}{分组变量名（字符串，列名，如 \code{"sex"}、\code{"location"} 等）。}

\item{location_name}{地区名。}

\item{measure_name}{指标名（如 \code{"Incidence"}）。}

\item{cause_name}{病因名。}

\item{rei_name}{可选，风险因素名。}

\item{sex_name}{性别(如 "Male", "Female", "Both"),如分组变量是"sex",则sex_name = c("Male", "Female", "Both")(任意组合)，如果分组变量不为"sex",则sex_name = "Male"(仅举一例)。}
}
\value{
返回 ggplot2 图形对象（双轴预测图）
}
\description{
用于对 \code{zstatsBAPC()} 函数的预测结果(包含 \code{ASR} 与 \code{all_age_projection})进行可视化，绘制双轴图。
}
