% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBDquantile.R
\name{zstatsquantile}
\alias{zstatsquantile}
\title{将连续型变量分为 n 个分位等级（默认五分位）}
\usage{
zstatsquantile(data, n = 5)
}
\arguments{
\item{data}{数值型向量，需进行分位划分的数据。}

\item{n}{分组数，默认为 5（即五分位）。}
}
\value{
一个因子变量，每个值对应其所属的分位组。
}
\description{
将连续变量按 quantile 划分为 n 个等级，如 "<20"、"20 to <40"。
}
