% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test2.R
\docType{package}
\name{test2}
\alias{test2}
\alias{test2-package}
\title{test2: A package to say hi}
\description{
The test2 package provides one function: \code{hello()}
}
\section{test2 functions}{

hello()
}

